block crosschain games

+++ backend/src/main/kotlin/com/retypeme/project/auth/SiweSession.kt
@@ -1,6 +1,6 @@
 package com.retypeme.project.auth
 
-class SiweSession(val address: String?, val chainId: Long?) {
+class SiweSession(val address: String?, val chainId: Int?) {
 
 }
